+++ src/models/driver_location.js
@@ -20,11 +20,11 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false
         },
         lat: {
-            type: DataTypes.DECIMAL,
+            type: DataTypes.DOUBLE,
             allowNull: false
         },
         long: {
-            type: DataTypes.DECIMAL,
+            type: DataTypes.DOUBLE,
             allowNull: false
         }
     }, {
